Met MVP. Trying to get an edit button working.

+++ src/scripts/deleteInterests.js
@@ -1,15 +1,22 @@
 import API from "./dataManager"
+import loadInterests from "./populateInterest";
 
 const deleteInterest = () => {
     document.querySelector("#populateInterests").addEventListener("click", () => {
         if (event.target.id.startsWith("delete")) {
             let id = event.target.id.split("--")[1]
 
-            return API.delete(id)
+            let check = confirm("Are you sure you want to delete this interest?")
+            if (check === true) {
+                return API.delete(id)
                 .then( () => {
-                    document.querySelector("#populateInterests").innerHTML = ""
-                    return
+                    loadInterests()
                 })
+            } else {
+                loadInterests()
+            }
+
+
         }
     })
 }
